+++ app/Http/Controllers/Admin/CategoryController.php
@@ -5,19 +5,25 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Models\Category;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class CategoryController extends Controller
 {
     public function index(){
+        if(!Gate::allows('categories')){
+            return view('admin.errors.notAllowed');
+        }
         $categories = Category::all();
         return view('admin.categories.index',compact('categories'));
     }//end of index
 
     public function create(){
+        if(!Gate::allows('categories.create')){
+            return view('admin.errors.notAllowed');
+        }
         $categories = Category::select('id', 'parent_id', 'name')->get();
         return view('admin.categories.create',compact('categories'));
     }//end of create
-
 
     public function store(Request $request, Category $category)
     {
@@ -41,7 +47,9 @@ class CategoryController extends Controller
     
     public function edit($id)
     {
-
+        if(!Gate::allows('categories.edit')){
+            return view('admin.errors.notAllowed');
+        }
         $category = Category::findOrFail($id);
         $categories = Category::select('id', 'parent_id', 'name')->get();
         if (!$category)
@@ -71,7 +79,9 @@ class CategoryController extends Controller
 
     public function destroy($id)
     {
-       
+        if(!Gate::allows('categories.destroy')){
+            return view('admin.errors.notAllowed');
+        }
         $category = Category::findOrFail($id);
         $category->delete();
         Toastr()->success('تم حذف القسم بنجاح');
